feat: REFACTORING STARTS AFTER THIS.
Learnt some lessons. In movement, combat and other player-centric
systems, a large part is a requirement for player data. Which
currently isn't present like how I want it to be.

Which is why I've consulted and figured out that a top-down approach
is better than a bottom-up. Systems that rely on other systems, should be
built after the required systems are built in this case specifically.

So I'll be going over and creating the player data system, inventory, etc.
All the backend logic, server-sided effects, etc. BEFORE getting started on
what the client needs